535. Encode and Decode TinyURL
